Split up handlers and server files
